Update to example code, added ADC functionality using JSW_ADC files.
I was able to read a value of 2728 from the FIFO register. Remembering that our ADC is 12 bits, in unsigned terms we were reading a voltage of approx. 2728/4095 = 0.66617 --> 0.066617*3.3V = 2.198V. Then using the ratio from our voltage divider (10k and 100k resistor) 2.198V * 11 = 24.122V. The actual measured value of the battery using a voltmeter was 24.3V. So less than 1% error.